requirements.txt
redis==4.4.0
psycopg2-binary==2.9.5
pydantic==1.10.3
boto3==1.26.40 
sendgrid==6.9.7